+++ tests/Jroc.Tests/BinaryOperator/GeneratorTests.cs
@@ -106,6 +106,9 @@ namespace Jroc.Tests.BinaryOperator
         public Task BinaryOperator_MulNumberNumber() { var testName = nameof(BinaryOperator_MulNumberNumber); return GenerateTest(testName); }
 
         [Fact]
+        public Task BinaryOperator_MulIndexedNumericCandidate() { var testName = nameof(BinaryOperator_MulIndexedNumericCandidate); return GenerateTest(testName); }
+
+        [Fact]
         public Task BinaryOperator_MulObjectObject() { var testName = nameof(BinaryOperator_MulObjectObject); return GenerateTest(testName); }
 
         [Fact]